feat(politeiagui-scripts): split atart and builf scripts for apps and plugins
