+++ plugins-structure/packages/politeiagui-scripts/config/webpack/webpack.config.js
@@ -4,18 +4,20 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 
 const appDirectory = fs.realpathSync(process.cwd());
 const resolveApp = (relativePath) => path.resolve(appDirectory, relativePath);
-const plugins = (webpackEnv) =>
-  webpackEnv === "development"
+const plugins = (isEnvDevelopment, isApp) =>
+  isApp
     ? [
         new HtmlWebpackPlugin({
           template: resolveApp("./src/public/index.html")
         })
       ]
-    : [
+    : isEnvDevelopment
+    ? [
         new HtmlWebpackPlugin({
-          template: resolveApp("./src/public/index.html")
+          template: resolveApp("./src/dev/index.html")
         })
-      ];
+      ]
+    : [];
 
 const pkgName = require(resolveApp("./package.json")).name;
 
@@ -67,16 +69,18 @@ const svgRules = [
   }
 ];
 
-module.exports = function (webpackEnv) {
+module.exports = function (webpackEnv = "production", type = "app") {
+  const isApp = type === "app";
   const isEnvDevelopment = webpackEnv === "development";
-  const isEnvProduction = webpackEnv === "production";
 
   return {
-    entry: isEnvDevelopment
+    entry: isApp
       ? resolveApp("./src/index.js")
+      : isEnvDevelopment
+      ? resolveApp("./src/dev/index.js")
       : resolveApp("./src/index.js"),
     target: ["browserslist"],
-    bail: isEnvProduction,
+    bail: !isEnvDevelopment,
     devtool: isEnvDevelopment ? "cheap-module-source-map" : "source-map",
     mode: isEnvDevelopment ? "development" : "production",
     output: {
@@ -97,6 +101,6 @@ module.exports = function (webpackEnv) {
         stream: require.resolve("stream-browserify")
       }
     },
-    plugins: plugins(webpackEnv)
+    plugins: plugins(isEnvDevelopment, isApp)
   };
 };
